% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS06.R
\docType{data}
\name{PS06}
\alias{PS06}
\title{Medicaid Expenditure 2019}
\format{
A data frame with 51 rows and 4 variables:
\describe{
\item{STATEFP}{STATEFP}
\item{TtlMedExpC}{Total medicaid spending in character format}
\item{TtlMedExpN}{Total medicaid spending in numeric format}
\item{Year}{year of the data}
}
}
\source{
\url{https://www.kff.org/medicaid/state-indicator/total-medicaid-spending/?}
}
\usage{
PS06
}
\description{
A dataset containing medicaid expenditure data at the state level for 2019.
}
\keyword{datasets}
